+++ examples/calc.php
@@ -1,7 +1,7 @@
 <?php
 
 use VovanVE\parser\common\Token;
-use VovanVE\parser\common\TreeNodeInterface;
+use VovanVE\parser\common\TreeNodeInterface as INode;
 use VovanVE\parser\grammar\Grammar;
 use VovanVE\parser\LexerBuilder;
 use VovanVE\parser\Parser;
@@ -33,30 +33,27 @@ $lexer = (new LexerBuilder)
 
 $actions = [
     'int' => function (Token $t) {
-        return (int) $t->getContent();
+        return (int)$t->getContent();
     },
-    'Value' => function ($v, TreeNodeInterface $n) {
-        return $n->made();
-    },
-    'Value(neg)' => function ($v, TreeNodeInterface $n) {
+
+    'Value' => Parser::ACTION_BUBBLE_THE_ONLY,
+    'Value(neg)' => function ($v, INode $n) {
         return -$n->made();
     },
-    'Product(V)' => function ($p, TreeNodeInterface $v) {
-        return $v->made();
-    },
-    'Product(mul)' => function ($p, TreeNodeInterface $a, TreeNodeInterface $b) {
+
+    'Product(V)' => Parser::ACTION_BUBBLE_THE_ONLY,
+    'Product(mul)' => function ($p, INode $a, INode $b) {
         return $a->made() * $b->made();
     },
-    'Product(div)' => function ($p, TreeNodeInterface $a, TreeNodeInterface $b) {
+    'Product(div)' => function ($p, INode $a, INode $b) {
         return $a->made() / $b->made();
     },
-    'Sum(P)' => function ($s, TreeNodeInterface $p) {
-        return $p->made();
-    },
-    'Sum(add)' => function ($s, TreeNodeInterface $a, TreeNodeInterface $b) {
+
+    'Sum(P)' => Parser::ACTION_BUBBLE_THE_ONLY,
+    'Sum(add)' => function ($s, INode $a, INode $b) {
         return $a->made() + $b->made();
     },
-    'Sum(sub)' => function ($s, TreeNodeInterface $a, TreeNodeInterface $b) {
+    'Sum(sub)' => function ($s, INode $a, INode $b) {
         return $a->made() - $b->made();
     },
 ];